@@ -9,16 +9,19 @@
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^8 || ^7.4",
+        "ext-json": "*",
         "api-clients/rx": "^2.2",
+        "evenement/evenement": "^3.0",
         "react/promise": "^2.7",
         "react/stream": "^1.1.0",
-        "reactivex/rxphp": "^2.0"
+        "reactivex/rxphp": "^2.0",
+        "thecodingmachine/safe": "^1.3"
     },
     "require-dev": {
-        "react/promise-stream": "^1.1",
-        "wyrihaximus/async-test-utilities": "^1.1",
-        "wyrihaximus/ticking-promise": "^1.6"
+        "react/promise-stream": "^1.2",
+        "wyrihaximus/async-test-utilities": "^4.2",
+        "wyrihaximus/ticking-promise": "^3"
     },
     "autoload": {
         "psr-4": {
@@ -31,62 +34,24 @@
         }
     },
     "config": {
-        "sort-packages": true,
+        "allow-plugins": {
+            "composer/package-versions-deprecated": true,
+            "infection/extension-installer": true,
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "icanhazstring/composer-unused": true,
+            "ergebnis/composer-normalize": true
+        },
         "platform": {
-            "php": "7.2"
-        }
+            "php": "7.4.7"
+        },
+        "sort-packages": true
     },
     "scripts": {
-        "ensure-installed": "composer install --ansi -n -q",
-        "cs": [
-            "@ensure-installed",
-            "php-cs-fixer fix --config=.php_cs --ansi --dry-run --diff --verbose --allow-risky=yes --show-progress=estimating"
+        "post-install-cmd": [
+            "composer normalize"
         ],
-        "cs-fix": [
-            "@ensure-installed",
-            "php-cs-fixer fix --config=.php_cs --ansi --verbose --allow-risky=yes --show-progress=estimating"
-        ],
-        "unit": [
-            "@ensure-installed",
-            "phpunit --colors=always -c phpunit.xml.dist"
-        ],
-        "unit-coverage": [
-            "@ensure-installed",
-            "phpunit --colors=always -c phpunit.xml.dist --coverage-text --coverage-html covHtml --coverage-clover ./build/logs/clover.xml"
-        ],
-        "lint-php": [
-            "@ensure-installed",
-            "parallel-lint --exclude vendor ."
-        ],
-        "qa-all": [
-            "@lint-php",
-            "@cs",
-            "@unit"
-        ],
-        "qa-all-coverage": [
-            "@lint-php",
-            "@cs",
-            "@unit-coverage"
-        ],
-        "qa-windows": [
-            "@lint-php",
-            "@cs",
-            "@unit"
-        ],
-        "qa-ci": [
-            "@unit"
-        ],
-        "qa-ci-extended": [
-            "@qa-all-coverage"
-        ],
-        "qa-ci-windows": [
-            "@qa-windows"
-        ],
-        "qa-contrib": [
-            "@qa-all"
-        ],
-        "ci-coverage": [
-            "if [ -f ./build/logs/clover.xml ]; then wget https://scrutinizer-ci.com/ocular.phar && php ocular.phar code-coverage:upload --format=php-clover ./build/logs/clover.xml; fi"
+        "post-update-cmd": [
+            "composer normalize"
         ]
     }
 }
