@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace WyriHaximus\React\Stream\Json;
 
@@ -6,10 +8,6 @@ use React\Stream\ReadableStreamInterface;
 
 final class BufferingReadableStream extends AbstractBufferingStream
 {
-    /**
-     * BufferingStream constructor.
-     * @param ReadableStreamInterface $stream
-     */
     public function __construct(ReadableStreamInterface $stream)
     {
         $this->setUp($stream);
