@@ -1,12 +1,15 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace WyriHaximus\React\Tests\Stream\Json;
 
-use function ApiClients\Tools\Rx\observableFromArray;
-use React\EventLoop\StreamSelectLoop;
-use function React\Promise\Stream\buffer;
+use React\EventLoop\Loop;
 use WyriHaximus\AsyncTestUtilities\AsyncTestCase;
 use WyriHaximus\React\Stream\Json\JsonStreamFactory;
+
+use function ApiClients\Tools\Rx\observableFromArray;
+use function React\Promise\Stream\buffer;
 
 /**
  * @internal
@@ -18,18 +21,17 @@ final class JsonStreamFactoryTest extends AsyncTestCase
      */
     public function createFromArray(): void
     {
-        $loop = new StreamSelectLoop();
         $array = [
             'cuvee',
             'buffalo',
         ];
 
         $stream = JsonStreamFactory::createFromArray($array);
-        $loop->futureTick(function () use ($stream): void {
+        Loop::futureTick(static function () use ($stream): void {
             $stream->resume();
         });
 
-        $json = $this->await(buffer($stream), $loop);
+        $json = $this->await(buffer($stream));
         self::assertSame('["cuvee","buffalo"]', $json);
     }
 
@@ -38,18 +40,17 @@ final class JsonStreamFactoryTest extends AsyncTestCase
      */
     public function createFromObservavle(): void
     {
-        $loop = new StreamSelectLoop();
         $array = [
             'cuvee',
             'buffalo',
         ];
 
         $stream = JsonStreamFactory::createFromObservable(observableFromArray($array));
-        $loop->futureTick(function () use ($stream): void {
+        Loop::futureTick(static function () use ($stream): void {
             $stream->resume();
         });
 
-        $json = $this->await(buffer($stream), $loop);
+        $json = $this->await(buffer($stream));
         self::assertSame('["cuvee","buffalo"]', $json);
     }
 }
