@@ -1,488 +1,540 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace WyriHaximus\React\Tests\Stream\Json;
 
-use function Clue\React\Block\await;
-use PHPUnit\Framework\TestCase;
-use React\EventLoop\Factory;
-use React\EventLoop\LoopInterface;
-use function React\Promise\all;
+use React\EventLoop\Loop;
 use React\Promise\Deferred;
-use function React\Promise\resolve;
-use function React\Promise\Stream\buffer;
 use React\Stream\ThroughStream;
 use Rx\Observable;
 use Rx\ObservableInterface;
 use Rx\Subject\Subject;
+use WyriHaximus\AsyncTestUtilities\AsyncTestCase;
 use WyriHaximus\React\Stream\Json\JsonStream;
+
+use function json_decode;
+use function json_last_error;
+use function range;
+use function React\Promise\all;
+use function React\Promise\resolve;
+use function React\Promise\Stream\buffer;
 use function WyriHaximus\React\timedPromise;
+
+use const JSON_ERROR_NONE;
+use const JSON_ERROR_SYNTAX;
 
 /**
  * @internal
  */
-final class JsonStreamTest extends TestCase
+final class JsonStreamTest extends AsyncTestCase
 {
-    public function provideJSONs()
+    /**
+     * @return iterable<array<callable>>
+     */
+    public function provideJSONs(): iterable
     {
-        yield [function () {
-            $input = [
-                'promise' => resolve('w00t'),
-            ];
+        yield [
+            static function (): array {
+                $input = [
+                    'promise' => resolve('w00t'),
+                ];
 
-            return [$input, '{"promise":"w00t"}'];
-        }];
+                return [$input, '{"promise":"w00t"}'];
+            },
+        ];
 
-        yield [function () {
-            $input = [
-                'promise' => all([resolve('w00t'),resolve('w00t'),resolve('1337'),resolve('train')]),
-            ];
+        yield [
+            static function (): array {
+                $input = [
+                    'promise' => all([resolve('w00t'), resolve('w00t'), resolve('1337'), resolve('train')]),
+                ];
 
-            return [$input, '{"promise":["w00t","w00t","1337","train"]}'];
-        }];
+                return [$input, '{"promise":["w00t","w00t","1337","train"]}'];
+            },
+        ];
 
-        yield [function () {
-            $input = [
-                'a',
-                'a',
-            ];
+        yield [
+            static function (): array {
+                $input = [
+                    'a',
+                    'a',
+                ];
 
-            return [$input, '["a","a"]'];
-        }];
+                return [$input, '["a","a"]'];
+            },
+        ];
 
-        yield [function () {
-            $input = [
-                'a' => 'b',
-                'c' => 'd',
-            ];
+        yield [
+            static function (): array {
+                $input = [
+                    'a' => 'b',
+                    'c' => 'd',
+                ];
 
-            return [$input, '{"a":"b","c":"d"}'];
-        }];
+                return [$input, '{"a":"b","c":"d"}'];
+            },
+        ];
 
-        yield [function () {
-            $input = [
-                'a' => resolve('b'),
-                'c' => resolve('d'),
-            ];
+        yield [
+            static function (): array {
+                $input = [
+                    'a' => resolve('b'),
+                    'c' => resolve('d'),
+                ];
 
-            return [$input, '{"a":"b","c":"d"}'];
-        }];
+                return [$input, '{"a":"b","c":"d"}'];
+            },
+        ];
 
-        yield [function () {
-            $input = [
-                'a' => resolve(['b']),
-                'c' => resolve(['d']),
-            ];
+        yield [
+            static function (): array {
+                $input = [
+                    'a' => resolve(['b']),
+                    'c' => resolve(['d']),
+                ];
 
-            return [$input, '{"a":["b"],"c":["d"]}'];
-        }];
+                return [$input, '{"a":["b"],"c":["d"]}'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $stream = new ThroughStream();
+        yield [
+            static function (): array {
+                $stream = new ThroughStream();
 
-            $input = [
-                'river' => $stream,
-            ];
+                $input = ['river' => $stream];
 
-            $loop->addTimer(0.05, function () use ($stream): void {
-                $stream->end('song');
-            });
+                Loop::addTimer(0.05, static function () use ($stream): void {
+                    $stream->end('song');
+                });
 
-            return [$input, '{"river":"song"}'];
-        }];
+                    return [$input, '{"river":"song"}'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $streamA = new ThroughStream();
-            $streamB = new ThroughStream();
+        yield [
+            static function (): array {
+                $streamA = new ThroughStream();
+                $streamB = new ThroughStream();
 
-            $input = [
-                'river' => $streamA,
-                'melody' => $streamB,
-            ];
+                $input = [
+                    'river' => $streamA,
+                    'melody' => $streamB,
+                ];
 
-            $loop->addTimer(0.1, function () use ($streamA): void {
-                $streamA->end('song');
-            });
+                Loop::addTimer(0.1, static function () use ($streamA): void {
+                    $streamA->end('song');
+                });
 
-            $loop->addTimer(0.05, function () use ($streamB): void {
-                $streamB->end('by the pond');
-            });
+                    Loop::addTimer(0.05, static function () use ($streamB): void {
+                        $streamB->end('by the pond');
+                    });
 
-            return [$input, '{"river":"song","melody":"by the pond"}'];
-        }];
+                    return [$input, '{"river":"song","melody":"by the pond"}'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $streamA = new ThroughStream();
-            $streamB = new ThroughStream();
-            $deferred = new Deferred();
+        yield [
+            static function (): array {
+                $streamA  = new ThroughStream();
+                $streamB  = new ThroughStream();
+                $deferred = new Deferred();
 
-            $input = [
-                'river' => $streamA,
-                'melody' => $streamB,
-                'from' => $deferred->promise(),
-            ];
+                $input = [
+                    'river' => $streamA,
+                    'melody' => $streamB,
+                    'from' => $deferred->promise(),
+                ];
 
-            $loop->addTimer(0.1, function () use ($streamA): void {
-                $streamA->end('song');
-            });
+                Loop::addTimer(0.1, static function () use ($streamA): void {
+                    $streamA->end('song');
+                });
 
-            $loop->addTimer(0.05, function () use ($streamB): void {
-                $streamB->end('by the pond');
-            });
+                    Loop::addTimer(0.05, static function () use ($streamB): void {
+                        $streamB->end('by the pond');
+                    });
 
-            $loop->addTimer(0.01, function () use ($deferred): void {
-                $deferred->resolve('the vortex');
-            });
+                    Loop::addTimer(0.01, static function () use ($deferred): void {
+                        $deferred->resolve('the vortex');
+                    });
 
-            return [$input, '{"river":"song","melody":"by the pond","from":"the vortex"}'];
-        }];
+                    return [$input, '{"river":"song","melody":"by the pond","from":"the vortex"}'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $streamA = new ThroughStream();
-            $streamB = new ThroughStream();
-            $streamCenturion = new ThroughStream();
-            $deferred = new Deferred();
-            $jsonStream = new JsonStream();
+        yield [
+            static function (): array {
+                $streamA         = new ThroughStream();
+                $streamB         = new ThroughStream();
+                $streamCenturion = new ThroughStream();
+                $deferred        = new Deferred();
+                $jsonStream      = new JsonStream();
 
-            $input = [
-                'river' => $streamA,
-                'melody' => $streamB,
-                'from' => $deferred->promise(),
-                'vortex' => $jsonStream,
-            ];
+                $input = [
+                    'river' => $streamA,
+                    'melody' => $streamB,
+                    'from' => $deferred->promise(),
+                    'vortex' => $jsonStream,
+                ];
 
-            $loop->addTimer(0.1, function () use ($streamA): void {
-                $streamA->end('song');
-            });
+                Loop::addTimer(0.1, static function () use ($streamA): void {
+                    $streamA->end('song');
+                });
 
-            $loop->addTimer(0.05, function () use ($streamB): void {
-                $streamB->end('by the pond');
-            });
+                    Loop::addTimer(0.05, static function () use ($streamB): void {
+                        $streamB->end('by the pond');
+                    });
 
-            $loop->addTimer(0.01, function () use ($deferred): void {
-                $deferred->resolve('the vortex');
-            });
+                    Loop::addTimer(0.01, static function () use ($deferred): void {
+                        $deferred->resolve('the vortex');
+                    });
 
-            $loop->addTimer(0.05, function () use ($jsonStream, $streamCenturion): void {
-                $jsonStream->end([
-                    'ponds' => resolve([
-                        'f' => resolve(resolve(resolve('the girl who waited'))),
-                        'm' => resolve($streamCenturion),
-                    ]),
-                ]);
-            });
+                    Loop::addTimer(0.05, static function () use ($jsonStream, $streamCenturion): void {
+                        $jsonStream->end([
+                            'ponds' => resolve([
+                                'f' => resolve(resolve(resolve('the girl who waited'))),
+                                'm' => resolve($streamCenturion),
+                            ]),
+                        ]);
+                    });
 
-            $loop->addTimer(0.1, function () use ($streamCenturion): void {
-                $streamCenturion->end('the last centurion');
-            });
+                    Loop::addTimer(0.1, static function () use ($streamCenturion): void {
+                        $streamCenturion->end('the last centurion');
+                    });
 
-            return [$input, '{"river":"song","melody":"by the pond","from":"the vortex","vortex":{"ponds":{"f":"the girl who waited","m":"the last centurion"}}}'];
-        }];
+                    return [$input, '{"river":"song","melody":"by the pond","from":"the vortex","vortex":{"ponds":{"f":"the girl who waited","m":"the last centurion"}}}'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $streamA = new ThroughStream();
-            $streamB = new ThroughStream();
-            $streamCenturion = new ThroughStream();
-            $deferred = new Deferred();
-            $jsonStream = new JsonStream();
+        yield [
+            static function (): array {
+                $streamA         = new ThroughStream();
+                $streamB         = new ThroughStream();
+                $streamCenturion = new ThroughStream();
+                $deferred        = new Deferred();
+                $jsonStream      = new JsonStream();
 
-            $input = [
-                'river' => $streamA,
-                'melody' => $streamB,
-                'from' => $deferred->promise(),
-                'vortex' => $jsonStream,
-                'timestream' => Observable::fromArray(['don\'t','blink']),
-            ];
+                $input = [
+                    'river' => $streamA,
+                    'melody' => $streamB,
+                    'from' => $deferred->promise(),
+                    'vortex' => $jsonStream,
+                    'timestream' => Observable::fromArray(['don\'t', 'blink']),
+                ];
 
-            $loop->addTimer(0.1, function () use ($streamA): void {
-                $streamA->end('song');
-            });
+                Loop::addTimer(0.1, static function () use ($streamA): void {
+                    $streamA->end('song');
+                });
 
-            $loop->addTimer(0.05, function () use ($streamB): void {
-                $streamB->end('by the pond');
-            });
+                    Loop::addTimer(0.05, static function () use ($streamB): void {
+                        $streamB->end('by the pond');
+                    });
 
-            $loop->addTimer(0.01, function () use ($deferred): void {
-                $deferred->resolve('the vortex');
-            });
+                    Loop::addTimer(0.01, static function () use ($deferred): void {
+                        $deferred->resolve('the vortex');
+                    });
 
-            $loop->addTimer(0.05, function () use ($jsonStream, $streamCenturion): void {
-                $jsonStream->end([
-                    'ponds' => resolve([
-                        'f' => resolve(resolve(resolve('the girl who waited'))),
-                        'm' => resolve($streamCenturion),
-                    ]),
-                ]);
-            });
+                    Loop::addTimer(0.05, static function () use ($jsonStream, $streamCenturion): void {
+                        $jsonStream->end([
+                            'ponds' => resolve([
+                                'f' => resolve(resolve(resolve('the girl who waited'))),
+                                'm' => resolve($streamCenturion),
+                            ]),
+                        ]);
+                    });
 
-            $loop->addTimer(0.1, function () use ($streamCenturion): void {
-                $streamCenturion->end('the last centurion');
-            });
+                    Loop::addTimer(0.1, static function () use ($streamCenturion): void {
+                        $streamCenturion->end('the last centurion');
+                    });
 
-            return [$input, '{"river":"song","melody":"by the pond","from":"the vortex","vortex":{"ponds":{"f":"the girl who waited","m":"the last centurion"}},"timestream":["don\u0027t","blink"]}'];
-        }];
+                    return [$input, '{"river":"song","melody":"by the pond","from":"the vortex","vortex":{"ponds":{"f":"the girl who waited","m":"the last centurion"}},"timestream":["don\u0027t","blink"]}'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $streamA = new ThroughStream();
-            $streamB = new ThroughStream();
-            $streamCenturion = new ThroughStream();
-            $deferred = new Deferred();
-            $jsonStream = new JsonStream();
-            $subject = new Subject();
+        yield [
+            static function (): array {
+                $streamA         = new ThroughStream();
+                $streamB         = new ThroughStream();
+                $streamCenturion = new ThroughStream();
+                $deferred        = new Deferred();
+                $jsonStream      = new JsonStream();
+                $subject         = new Subject();
 
-            $input = [
-                'timestream' => $subject,
-                'river' => $streamA,
-                'melody' => $streamB,
-                'from' => $deferred->promise(),
-                'vortex' => $jsonStream,
-            ];
+                $input = [
+                    'timestream' => $subject,
+                    'river' => $streamA,
+                    'melody' => $streamB,
+                    'from' => $deferred->promise(),
+                    'vortex' => $jsonStream,
+                ];
 
-            $loop->addTimer(0.1, function () use ($streamA): void {
-                $streamA->end('song');
-            });
+                Loop::addTimer(0.1, static function () use ($streamA): void {
+                    $streamA->end('song');
+                });
 
-            $loop->addTimer(0.05, function () use ($streamB): void {
-                $streamB->end('by the pond');
-            });
+                    Loop::addTimer(0.05, static function () use ($streamB): void {
+                        $streamB->end('by the pond');
+                    });
 
-            $loop->addTimer(0.01, function () use ($deferred): void {
-                $deferred->resolve('the vortex');
-            });
+                    Loop::addTimer(0.01, static function () use ($deferred): void {
+                        $deferred->resolve('the vortex');
+                    });
 
-            $loop->addTimer(0.05, function () use ($jsonStream, $streamCenturion): void {
-                $jsonStream->end([
-                    'ponds' => resolve([
-                        'f' => resolve(resolve(resolve('the girl who waited'))),
-                        'm' => resolve($streamCenturion),
-                    ]),
-                ]);
-            });
+                    Loop::addTimer(0.05, static function () use ($jsonStream, $streamCenturion): void {
+                        $jsonStream->end([
+                            'ponds' => resolve([
+                                'f' => resolve(resolve(resolve('the girl who waited'))),
+                                'm' => resolve($streamCenturion),
+                            ]),
+                        ]);
+                    });
 
-            $loop->addTimer(0.1, function () use ($streamCenturion): void {
-                $streamCenturion->end('the last centurion');
-            });
+                    Loop::addTimer(0.1, static function () use ($streamCenturion): void {
+                        $streamCenturion->end('the last centurion');
+                    });
 
-            $loop->addTimer(0.2, function () use ($subject): void {
-                $subject->onNext('don\'t');
-            });
+                    Loop::addTimer(0.2, static function () use ($subject): void {
+                        $subject->onNext('don\'t');
+                    });
 
-            $loop->addTimer(0.7, function () use ($subject): void {
-                $subject->onNext('blink');
-            });
+                    Loop::addTimer(0.7, static function () use ($subject): void {
+                        $subject->onNext('blink');
+                    });
 
-            $loop->addTimer(0.9, function () use ($subject): void {
-                $subject->onCompleted();
-            });
+                    Loop::addTimer(0.9, static function () use ($subject): void {
+                        $subject->onCompleted();
+                    });
 
-            return [$input, '{"timestream":["don\u0027t","blink"],"river":"song","melody":"by the pond","from":"the vortex","vortex":{"ponds":{"f":"the girl who waited","m":"the last centurion"}}}'];
-        }];
+                    return [$input, '{"timestream":["don\u0027t","blink"],"river":"song","melody":"by the pond","from":"the vortex","vortex":{"ponds":{"f":"the girl who waited","m":"the last centurion"}}}'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $stream = new ThroughStream();
+        yield [
+            static function (): array {
+                $stream = new ThroughStream();
 
-            $input = [
-                'a' => resolve([
-                    'b' => resolve([
-                        'c' => resolve([
-                            'd' => resolve(resolve(resolve(resolve(resolve($stream))))),
+                $input = [
+                    'a' => resolve([
+                        'b' => resolve([
+                            'c' => resolve([
+                                'd' => resolve(resolve(resolve(resolve(resolve($stream))))),
+                            ]),
                         ]),
                     ]),
-                ]),
-            ];
+                ];
 
-            $loop->addTimer(0.1, function () use ($stream): void {
-                $stream->end('e');
-            });
+                Loop::addTimer(0.1, static function () use ($stream): void {
+                    $stream->end('e');
+                });
 
-            return [$input, '{"a":{"b":{"c":{"d":"e"}}}}'];
-        }];
+                    return [$input, '{"a":{"b":{"c":{"d":"e"}}}}'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $stream = new ThroughStream();
+        yield [
+            static function (): array {
+                $stream = new ThroughStream();
 
-            $input = [
-                'a' => resolve([
-                    'b' => [
-                        'c' => resolve([
-                            'd' => resolve(resolve(resolve(resolve(resolve($stream))))),
-                        ]),
+                $input = [
+                    'a' => resolve([
+                        'b' => [
+                            'c' => resolve([
+                                'd' => resolve(resolve(resolve(resolve(resolve($stream))))),
+                            ]),
+                        ],
+                    ]),
+                ];
+
+                Loop::addTimer(0.1, static function () use ($stream): void {
+                    $stream->end('e');
+                });
+
+                    return [$input, '{"a":{"b":{"c":{"d":"e"}}}}'];
+            },
+        ];
+
+        yield [
+            static function (): array {
+                $input = range(1, 10);
+
+                return [$input, '[1,2,3,4,5,6,7,8,9,10]'];
+            },
+        ];
+
+        yield [
+            static function (): array {
+                $input = [
+                    resolve(1),
+                    resolve(2),
+                    resolve(3),
+                    resolve(4),
+                    resolve(5),
+                    resolve(6),
+                    resolve(7),
+                    resolve(8),
+                    resolve(9),
+                    resolve(10),
+                ];
+
+                return [$input, '[1,2,3,4,5,6,7,8,9,10]'];
+            },
+        ];
+
+        yield [
+            static function (): array {
+                $input = [
+                    ['foo' => 'bar'],
+                    ['bar' => 'foo'],
+                ];
+
+                return [$input, '[{"foo":"bar"},{"bar":"foo"}]'];
+            },
+        ];
+
+        yield [
+            static function (): array {
+                $stream = new ThroughStream();
+
+                $input = [
+                    [
+                        'foo' => resolve('bar'),
                     ],
-                ]),
-            ];
+                    resolve(['bar' => $stream]),
+                ];
 
-            $loop->addTimer(0.1, function () use ($stream): void {
-                $stream->end('e');
-            });
+                Loop::addTimer(0.1, static function () use ($stream): void {
+                    $stream->end('foo');
+                });
 
-            return [$input, '{"a":{"b":{"c":{"d":"e"}}}}'];
-        }];
+                    return [$input, '[{"foo":"bar"},{"bar":"foo"}]'];
+            },
+        ];
 
-        yield [function () {
-            $input = \range(1, 10);
+        yield [
+            static function (): array {
+                $stream = new ThroughStream();
 
-            return [$input, '[1,2,3,4,5,6,7,8,9,10]'];
-        }];
+                $input = ['😱' => $stream];
 
-        yield [function () {
-            $input = [
-                resolve(1),
-                resolve(2),
-                resolve(3),
-                resolve(4),
-                resolve(5),
-                resolve(6),
-                resolve(7),
-                resolve(8),
-                resolve(9),
-                resolve(10),
-            ];
+                Loop::addTimer(0.05, static function () use ($stream): void {
+                    $stream->end('😱');
+                });
 
-            return [$input, '[1,2,3,4,5,6,7,8,9,10]'];
-        }];
+                    return [$input, '{"\ud83d\ude31":"\ud83d\ude31"}'];
+            },
+        ];
 
-        yield [function () {
-            $input = [
-                [
-                    'foo' => 'bar',
-                ],
-                [
-                    'bar' => 'foo',
-                ],
-            ];
+        yield [
+            static function (): array {
+                $stream = new ThroughStream();
 
-            return [$input, '[{"foo":"bar"},{"bar":"foo"}]'];
-        }];
+                $input = ['😱' => $stream];
 
-        yield [function (LoopInterface $loop) {
-            $stream = new ThroughStream();
+                Loop::addTimer(0.05, static function () use ($stream): void {
+                    $stream->end('<\'&"&\'>');
+                });
 
-            $input = [
-                [
-                    'foo' => resolve('bar'),
-                ],
-                resolve([
-                    'bar' => $stream,
-                ]),
-            ];
+                    return [$input, '{"\ud83d\ude31":"\u003C\u0027\u0026\u0022\u0026\u0027\u003E"}'];
+            },
+        ];
 
-            $loop->addTimer(0.1, function () use ($stream): void {
-                $stream->end('foo');
-            });
+        yield [
+            static function (): array {
+                $input = [
+                    '😱' => Observable::fromArray(['foo', 'bar']),
+                ];
 
-            return [$input, '[{"foo":"bar"},{"bar":"foo"}]'];
-        }];
+                return [$input, '{"\ud83d\ude31":["foo","bar"]}'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $stream = new ThroughStream();
+        yield [
+            static function (): array {
+                $input = Observable::fromArray(['foo', 'bar']);
 
-            $input = [
-                '😱' => $stream,
-            ];
+                return [$input, '["foo","bar"]'];
+            },
+        ];
 
-            $loop->addTimer(0.05, function () use ($stream): void {
-                $stream->end('😱');
-            });
+        yield [
+            static function (): array {
+                $input = Observable::fromArray([resolve('foo'), timedPromise(0.3, 'bar')]);
 
-            return [$input, '{"\ud83d\ude31":"\ud83d\ude31"}'];
-        }];
+                return [$input, '["foo","bar"]'];
+            },
+        ];
 
-        yield [function (LoopInterface $loop) {
-            $stream = new ThroughStream();
+        yield [
+            static function (): array {
+                $streamA         = new ThroughStream();
+                $streamB         = new ThroughStream();
+                $streamCenturion = new ThroughStream();
+                $deferred        = new Deferred();
+                $jsonStream      = new JsonStream();
+                $subject         = new Subject();
 
-            $input = [
-                '😱' => $stream,
-            ];
-
-            $loop->addTimer(0.05, function () use ($stream): void {
-                $stream->end('<\'&"&\'>');
-            });
-
-            return [$input, '{"\ud83d\ude31":"\u003C\u0027\u0026\u0022\u0026\u0027\u003E"}'];
-        }];
-
-        yield [function (LoopInterface $loop) {
-            $input = [
-                '😱' => Observable::fromArray(['foo','bar']),
-            ];
-
-            return [$input, '{"\ud83d\ude31":["foo","bar"]}'];
-        }];
-
-        yield [function (LoopInterface $loop) {
-            $input = Observable::fromArray(['foo','bar']);
-
-            return [$input, '["foo","bar"]'];
-        }];
-
-        yield [function (LoopInterface $loop) {
-            $input = Observable::fromArray([resolve('foo'), timedPromise($loop, 0.3, 'bar')]);
-
-            return [$input, '["foo","bar"]'];
-        }];
-
-        yield [function (LoopInterface $loop) {
-            $streamA = new ThroughStream();
-            $streamB = new ThroughStream();
-            $streamCenturion = new ThroughStream();
-            $deferred = new Deferred();
-            $jsonStream = new JsonStream();
-            $subject = new Subject();
-
-            $input = Observable::fromArray([resolve('foo'), timedPromise($loop, 0.3, 'bar'), [
-                'timestream' => $subject,
-                'river' => $streamA,
-                'melody' => $streamB,
-                'from' => $deferred->promise(),
-                'vortex' => $jsonStream,
-            ]]);
-
-            $loop->addTimer(0.1, function () use ($streamA): void {
-                $streamA->end('song');
-            });
-
-            $loop->addTimer(0.05, function () use ($streamB): void {
-                $streamB->end('by the pond');
-            });
-
-            $loop->addTimer(0.01, function () use ($deferred): void {
-                $deferred->resolve('the vortex');
-            });
-
-            $loop->addTimer(0.05, function () use ($jsonStream, $streamCenturion): void {
-                $jsonStream->end([
-                    'ponds' => resolve([
-                        'f' => resolve(resolve(resolve('the girl who waited'))),
-                        'm' => resolve($streamCenturion),
-                    ]),
+                $input = Observable::fromArray([
+                    resolve('foo'),
+                    timedPromise(0.3, 'bar'),
+                    [
+                        'timestream' => $subject,
+                        'river' => $streamA,
+                        'melody' => $streamB,
+                        'from' => $deferred->promise(),
+                        'vortex' => $jsonStream,
+                    ],
                 ]);
-            });
 
-            $loop->addTimer(0.1, function () use ($streamCenturion): void {
-                $streamCenturion->end('the last centurion');
-            });
+                    Loop::addTimer(0.1, static function () use ($streamA): void {
+                        $streamA->end('song');
+                    });
 
-            $loop->addTimer(0.2, function () use ($subject): void {
-                $subject->onNext('don\'t');
-            });
+                    Loop::addTimer(0.05, static function () use ($streamB): void {
+                        $streamB->end('by the pond');
+                    });
 
-            $loop->addTimer(0.7, function () use ($subject): void {
-                $subject->onNext('blink');
-            });
+                    Loop::addTimer(0.01, static function () use ($deferred): void {
+                        $deferred->resolve('the vortex');
+                    });
 
-            $loop->addTimer(0.9, function () use ($subject): void {
-                $subject->onCompleted();
-            });
+                    Loop::addTimer(0.05, static function () use ($jsonStream, $streamCenturion): void {
+                        $jsonStream->end([
+                            'ponds' => resolve([
+                                'f' => resolve(resolve(resolve('the girl who waited'))),
+                                'm' => resolve($streamCenturion),
+                            ]),
+                        ]);
+                    });
 
-            return [$input, '["foo","bar",{"timestream":["don\u0027t","blink"],"river":"song","melody":"by the pond","from":"the vortex","vortex":{"ponds":{"f":"the girl who waited","m":"the last centurion"}}}]'];
-        }];
+                    Loop::addTimer(0.1, static function () use ($streamCenturion): void {
+                        $streamCenturion->end('the last centurion');
+                    });
 
-        yield [function (LoopInterface $loop) {
-            $input = [
-                'data' => Observable::fromArray([timedPromise($loop, 0.1, 'foo'), timedPromise($loop, 0.3, 'bar')]),
-            ];
+                    Loop::addTimer(0.2, static function () use ($subject): void {
+                        $subject->onNext('don\'t');
+                    });
 
-            return [$input, '{"data":["foo","bar"]}'];
-        }];
+                    Loop::addTimer(0.7, static function () use ($subject): void {
+                        $subject->onNext('blink');
+                    });
+
+                    Loop::addTimer(0.9, static function () use ($subject): void {
+                        $subject->onCompleted();
+                    });
+
+                    return [$input, '["foo","bar",{"timestream":["don\u0027t","blink"],"river":"song","melody":"by the pond","from":"the vortex","vortex":{"ponds":{"f":"the girl who waited","m":"the last centurion"}}}]'];
+            },
+        ];
+
+        yield [
+            static function (): array {
+                $input = [
+                    'data' => Observable::fromArray([timedPromise(0.1, 'foo'), timedPromise(0.3, 'bar')]),
+                ];
+
+                return [$input, '{"data":["foo","bar"]}'];
+            },
+        ];
     }
 
     /**
@@ -490,8 +542,7 @@ final class JsonStreamTest extends TestCase
      */
     public function testStream(callable $args): void
     {
-        $loop = Factory::create();
-        list($input, $output) = $args($loop);
+        [$input, $output] = $args();
 
         $throughStream = new ThroughStream();
 
@@ -499,243 +550,220 @@ final class JsonStreamTest extends TestCase
 
         $stream->pipe($throughStream);
 
-        $loop->addTimer(0.01, function () use ($stream, $input): void {
+        Loop::addTimer(0.01, static function () use ($stream, $input): void {
             if ($input instanceof ObservableInterface) {
                 $stream->writeObservable($input);
                 $stream->end();
 
                 return;
             }
+
             $stream->end($input);
         });
 
-        $buffer = await(buffer($throughStream), $loop, 6);
+        $buffer = $this->await(buffer($throughStream), 6);
 
         self::assertSame($output, $buffer);
-        $json = \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        $json = json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
         self::assertIsArray($json);
-        self::assertSame(\json_decode($output, true), $json);
+        self::assertSame(json_decode($output, true), $json); /** @phpstan-ignore-line */
     }
 
     public function testEncodeFlags(): void
     {
-        $loop = Factory::create();
-
         $stream = new JsonStream(0);
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->end(['<\'&"&\'>']);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('["<\'&\"&\'>"]', $buffer);
-        $json = \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        $json = json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
         self::assertSame(['<\'&"&\'>'], $json);
     }
 
     public function testObjectOrArrayObject(): void
     {
-        $loop = Factory::create();
-
         $stream = new JsonStream();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->writeValue(true);
             $stream->end([false]);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('{true,false}', $buffer);
-        \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_SYNTAX, \json_last_error());
+        json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_SYNTAX, json_last_error());
     }
 
     public function testObjectOrArrayArray(): void
     {
-        $loop = Factory::create();
-
         $stream = new JsonStream();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->end([true, false]);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('[true,false]', $buffer);
-        \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
     }
 
     public function testForceArray(): void
     {
-        $loop = Factory::create();
-
         $stream = JsonStream::createArray();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->end([true, false]);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('[true,false]', $buffer);
-        \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
     }
 
     public function testForceArrayWhileWeWriteAnObject(): void
     {
-        $loop = Factory::create();
-
         $stream = JsonStream::createArray();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->end(['a' => true, 'b' => false]);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('["a":true,"b":false]', $buffer);
-        \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_SYNTAX, \json_last_error());
+        json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_SYNTAX, json_last_error());
     }
 
     public function testForceObject(): void
     {
-        $loop = Factory::create();
-
         $stream = JsonStream::createObject();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->end(['a' => true, 'b' => false]);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('{"a":true,"b":false}', $buffer);
-        \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
     }
 
     public function testForceObjectWhileWeWriteAnArray(): void
     {
-        $loop = Factory::create();
-
         $stream = JsonStream::createObject();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->end([true, false]);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('{true,false}', $buffer);
-        \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_SYNTAX, \json_last_error());
+        json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_SYNTAX, json_last_error());
     }
 
     public function testDoubleKeys(): void
     {
-        $loop = Factory::create();
-
         $stream = JsonStream::createObject();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->write('a', true);
             $stream->write('a', false);
             $stream->end();
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('{"a":true,"a":false}', $buffer);
-        $json = \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        $json = json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
         self::assertSame(['a' => false], $json);
     }
 
     public function testNoMoreWriteAfterEnd(): void
     {
-        $loop = Factory::create();
-
         $stream = JsonStream::createObject();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->write('a', true);
             $stream->end();
             $stream->write('b', false);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('{"a":true}', $buffer);
-        $json = \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        $json = json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
         self::assertSame(['a' => true], $json);
     }
 
     public function testNoMoreWriteValueAfterEnd(): void
     {
-        $loop = Factory::create();
-
         $stream = JsonStream::createObject();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->write('a', true);
             $stream->end();
             $stream->writeValue(false);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('{"a":true}', $buffer);
-        $json = \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        $json = json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
         self::assertSame(['a' => true], $json);
     }
 
     public function testNoMoreWriteArrayAfterEnd(): void
     {
-        $loop = Factory::create();
-
         $stream = JsonStream::createObject();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->write('a', true);
             $stream->end();
-            $stream->writeArray([1,2,3,4,5,6,7,8,9,10]);
+            $stream->writeArray([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('{"a":true}', $buffer);
-        $json = \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        $json = json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
         self::assertSame(['a' => true], $json);
     }
 
     public function testNoMoreEndAfterEnd(): void
     {
-        $loop = Factory::create();
-
         $stream = JsonStream::createObject();
 
-        $loop->addTimer(0.01, function () use ($stream): void {
+        Loop::addTimer(0.01, static function () use ($stream): void {
             $stream->write('a', true);
             $stream->end();
-            $stream->end([1,2,3,4,5,6,7,8,9,10]);
+            $stream->end([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
         });
 
-        $buffer = await(buffer($stream), $loop, 6);
+        $buffer = $this->await(buffer($stream), 6);
 
         self::assertSame('{"a":true}', $buffer);
-        $json = \json_decode($buffer, true);
-        self::assertSame(\JSON_ERROR_NONE, \json_last_error());
+        $json = json_decode($buffer, true); /** @phpstan-ignore-line */
+        self::assertSame(JSON_ERROR_NONE, json_last_error());
         self::assertSame(['a' => true], $json);
     }
 
@@ -744,11 +772,11 @@ final class JsonStreamTest extends TestCase
         $stream = JsonStream::createObject();
 
         $shouldntBeCalledCount = 0;
-        $shouldntBeCalled = function () use (&$shouldntBeCalledCount): void {
+        $shouldntBeCalled      = static function () use (&$shouldntBeCalledCount): void {
             $shouldntBeCalledCount++;
         };
-        $shouldBeCalledCount = 0;
-        $shouldBeCalled = function () use (&$shouldBeCalledCount): void {
+        $shouldBeCalledCount   = 0;
+        $shouldBeCalled        = static function () use (&$shouldBeCalledCount): void {
             $shouldBeCalledCount++;
         };
 
@@ -772,16 +800,16 @@ final class JsonStreamTest extends TestCase
         $stream->resume();
 
         self::assertSame(0, $shouldntBeCalledCount);
-        self::assertSame(1, $shouldBeCalledCount);
+        self::assertSame(1, $shouldBeCalledCount); /** @phpstan-ignore-line */
 
         $stream->write('key', 'value');
 
         self::assertSame(0, $shouldntBeCalledCount);
-        self::assertSame(4, $shouldBeCalledCount);
+        self::assertSame(4, $shouldBeCalledCount); /** @phpstan-ignore-line */
 
         $stream->end();
 
         self::assertSame(0, $shouldntBeCalledCount);
-        self::assertSame(5, $shouldBeCalledCount);
+        self::assertSame(5, $shouldBeCalledCount); /** @phpstan-ignore-line */
     }
 }
